Technical/Update gem development dependencies
* Updated gem development dependencies
* Updated reek config
* Changed namespace for rspec helpers
