--- spec/support/helpers/dependency_helper.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module SmtpMock
-  module DependencyHelper
-    def compose_command(command_line_args)
-      SmtpMock::Dependency.compose_command(command_line_args)
-    end
-  end
-end
